Separate files in two modules: core and platform
The core module will contain the bulk of Glean.js logic. It will also
define the interfaces for the modules that should be implemented per platform.

The platform module will contain platform specific code. A different
module with the exact same interface should be implemented for each platform.

It will be up to our users to pass the platform specific code to Glean
on initialize. This commit also include the changes that add `platform`
as an initialization argument.